[logic][l]: - bitstore movedt separate module from app/packages/modules - tests moved to test/bitstore - imports and mocks updated appropriately refs #401
